improved and optimized mouse selection
