+++ src/config.rs
@@ -94,125 +94,6 @@ impl BootstrapConfig {
     }
 }
 
-/// Migrate notes and key from old storage path to new storage path
-pub fn migrate_storage(old_path: &PathBuf, new_path: &PathBuf) -> Result<MigrationResult> {
-    let old_notes_dir = old_path.join("notes");
-    let old_key_path = old_path.join("key.bin");
-    let old_settings_path = old_path.join("settings.json");
-    let old_keybinds_path = old_path.join("keybinds.toml");
-    let old_templates_dir = old_path.join("templates");
-
-    let new_notes_dir = new_path.join("notes");
-    let new_key_path = new_path.join("key.bin");
-    let new_settings_path = new_path.join("settings.json");
-    let new_keybinds_path = new_path.join("keybinds.toml");
-    let new_templates_dir = new_path.join("templates");
-
-    // Create new directories
-    fs::create_dir_all(&new_notes_dir).context("failed to create new notes directory")?;
-
-    let mut result = MigrationResult::default();
-
-    // Migrate key file
-    if old_key_path.exists() && !new_key_path.exists() {
-        fs::copy(&old_key_path, &new_key_path).context("failed to copy encryption key")?;
-        result.key_migrated = true;
-    }
-
-    // Migrate settings
-    if old_settings_path.exists() && !new_settings_path.exists() {
-        fs::copy(&old_settings_path, &new_settings_path).context("failed to copy settings")?;
-        result.settings_migrated = true;
-    }
-
-    // Migrate keybinds
-    if old_keybinds_path.exists() && !new_keybinds_path.exists() {
-        fs::copy(&old_keybinds_path, &new_keybinds_path).context("failed to copy keybinds")?;
-        result.keybinds_migrated = true;
-    }
-
-    // Migrate notes
-    if old_notes_dir.exists() {
-        for entry in fs::read_dir(&old_notes_dir).context("failed to read old notes directory")? {
-            let entry = entry.context("failed to read note entry")?;
-            let old_note_path = entry.path();
-            if let Some(filename) = old_note_path.file_name() {
-                let new_note_path = new_notes_dir.join(filename);
-                if !new_note_path.exists() {
-                    fs::copy(&old_note_path, &new_note_path).context("failed to copy note")?;
-                    result.notes_migrated += 1;
-                }
-            }
-        }
-    }
-
-    // Migrate templates
-    if old_templates_dir.exists() {
-        fs::create_dir_all(&new_templates_dir)
-            .context("failed to create new templates directory")?;
-        for entry in
-            fs::read_dir(&old_templates_dir).context("failed to read old templates directory")?
-        {
-            let entry = entry.context("failed to read template entry")?;
-            let old_template_path = entry.path();
-            if let Some(filename) = old_template_path.file_name() {
-                let new_template_path = new_templates_dir.join(filename);
-                if !new_template_path.exists() {
-                    fs::copy(&old_template_path, &new_template_path)
-                        .context("failed to copy template")?;
-                    result.templates_migrated += 1;
-                }
-            }
-        }
-    }
-
-    Ok(result)
-}
-
-#[derive(Debug, Default)]
-pub struct MigrationResult {
-    pub key_migrated: bool,
-    pub settings_migrated: bool,
-    pub keybinds_migrated: bool,
-    pub notes_migrated: usize,
-    pub templates_migrated: usize,
-}
-
-impl MigrationResult {
-    pub fn summary(&self) -> String {
-        let mut parts = Vec::new();
-        if self.key_migrated {
-            parts.push("encryption key".to_string());
-        }
-        if self.settings_migrated {
-            parts.push("settings".to_string());
-        }
-        if self.keybinds_migrated {
-            parts.push("keybinds".to_string());
-        }
-        if self.notes_migrated > 0 {
-            parts.push(format!("{} note(s)", self.notes_migrated));
-        }
-        if self.templates_migrated > 0 {
-            parts.push(format!("{} template(s)", self.templates_migrated));
-        }
-
-        if parts.is_empty() {
-            "No files migrated (destination already has data or source is empty)".to_string()
-        } else {
-            format!("Migrated: {}", parts.join(", "))
-        }
-    }
-
-    pub fn has_migrations(&self) -> bool {
-        self.key_migrated
-            || self.settings_migrated
-            || self.keybinds_migrated
-            || self.notes_migrated > 0
-            || self.templates_migrated > 0
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
